requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
fonttools==4.38.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==2.10.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.1.0
joblib==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
networkx==2.8.8
numpy==1.21.4
oauthlib==3.2.2
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.3
Pillow==9.3.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==0.23.0
scipy==1.4.1
setuptools-scm==7.0.5
Shapely==1.8.5.post1
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
